Use guard, move runtime_stats to metrics
