+++ src/coprocessor/util.rs
@@ -134,60 +134,6 @@ pub fn get_pk(col: &ColumnInfo, h: i64) -> Datum {
     }
 }
 
-use std::time::Duration;
-
-// RuntimeStats collects one executor's execution info.
-#[derive(Debug, Default)]
-pub struct RuntimeStats {
-    // executor consume time.
-    pub consume: Duration,
-    // executor's Next() called times.
-    pub count: i64,
-    // executor return row count.
-    pub rows: i64,
-}
-
-#[allow(dead_code)]
-impl RuntimeStats {
-    // Record records executor's execution.
-    pub fn record(&mut self, d: Duration, row_num: i64) {
-        self.consume += d;
-        self.count += 1;
-        self.rows += row_num;
-    }
-
-    // Record and set records executor's elapsed time and set row_num
-    pub fn record_and_set(&mut self, d: Duration, row_num: i64) {
-        self.consume += d;
-        self.count += 1;
-        self.rows = row_num;
-    }
-
-    // set_row_num sets the row num.
-    pub fn set_row_num(&mut self, row_num: i64) {
-        self.rows = row_num;
-    }
-
-    // string returns the information
-    pub fn string(&mut self) -> String {
-        format!(
-            "time:{}, loops:{}, rows:{}",
-            self.consume.as_nanos(),
-            self.count,
-            self.rows,
-        )
-    }
-
-    // initialize structure
-    pub fn empty() -> RuntimeStats {
-        RuntimeStats {
-            consume: Duration::new(0, 0),
-            count: 0,
-            rows: 0,
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -260,76 +206,5 @@ mod tests {
                 &[1, 2, 4, 0, 0, 0],
             ],
         );
-    }
-
-    use std::time::Duration;
-
-    #[test]
-    fn test_record() {
-        let mut run_time_stats = RuntimeStats {
-            consume: Duration::new(0, 1),
-            count: 2,
-            rows: 3,
-        };
-        let cases = vec![(1, 1, 2, 3, 4), (2, 2, 4, 4, 6)];
-
-        for (d, row_num, exp_time, exp_count, exp_rows) in cases {
-            run_time_stats.record(Duration::from_nanos(d), row_num);
-            assert_eq!(
-                run_time_stats.string(),
-                format!("time:{}, loops:{}, rows:{}", exp_time, exp_count, exp_rows),
-            );
-        }
-    }
-
-    #[test]
-    fn test_record_and_set() {
-        let mut run_time_stats = RuntimeStats {
-            consume: Duration::new(0, 1),
-            count: 2,
-            rows: 3,
-        };
-        let cases = vec![(1, 1, 2, 3, 1), (2, 2, 4, 4, 2)];
-
-        for (d, row_num, exp_time, exp_count, exp_rows) in cases {
-            run_time_stats.record_and_set(Duration::from_nanos(d), row_num);
-            assert_eq!(
-                run_time_stats.string(),
-                format!("time:{}, loops:{}, rows:{}", exp_time, exp_count, exp_rows),
-            );
-        }
-    }
-
-    #[test]
-    fn test_set_row_num() {
-        let mut run_time_stats = RuntimeStats {
-            consume: Duration::new(0, 1),
-            count: 2,
-            rows: 3,
-        };
-
-        let cases = vec![1, 2, 3];
-        for row_num in cases {
-            run_time_stats.set_row_num(row_num);
-            assert_eq!(
-                run_time_stats.string(),
-                format!("time:1, loops:2, rows:{}", row_num),
-            );
-        }
-    }
-
-    #[test]
-    fn test_string() {
-        let mut run_time_stats = RuntimeStats {
-            consume: Duration::new(0, 1),
-            count: 2,
-            rows: 3,
-        };
-        assert_eq!(run_time_stats.string(), "time:1, loops:2, rows:3");
-    }
-
-    #[test]
-    fn test_empty() {
-        assert_eq!(RuntimeStats::empty().string(), "time:0, loops:0, rows:0");
     }
 }
